requirements.txt
streamlit
langchain
langchain-groq
python-pptx
python-dotenv
